init: drop remapping
when the AT91 boots, the flash memory is mapped to address 0x0. The init
code remapped the SRAM to this memory and the linker script was
configured to use 0x0 for RAM. We don't have any reason to do this, so
delete this code. This will allow us to be more flexible in the linker
script.

The linker script is deleted since it is no longer valid and it was not
being used.